VC8 is a bit stricter on casting and equality...
git-svn-id: svn://falconpl.org/falcon/core/trunk@547 78e5f83a-d9f7-4e86-b854-dad15b5eb37a

+++ engine/core_module/table.cpp
@@ -339,7 +339,7 @@ static uint32 internal_col_pos( CoreTable *table, VMachine *vm, Item *i_column )
 
 static void internal_get_item( CoreTable *table, CoreArray *row, VMachine *vm, Item *i_column )
 {
-   if ( i_column->isInteger() && i_column->asInteger() == CoreTable::noitem )
+   if ( i_column->isInteger() && (((uint32)i_column->asInteger()) == CoreTable::noitem) )
    {
       vm->retval( row );
       return;
